Stop restoring apollo cache after the appsync store has been rehydrated
Constantly restoring apollo cache from store after writing causes issues
with optimistic responses. The patches are not pruned and ultimately the
appsync store does not end up with the correct data and the apollo cache
is 'corrupted'. Easily reproducible by creating a mutation with a
pessimistic boolean flag on optimistic response, no update function (as
to not update an existing query) and no running subscription.