Stop persisting recordingCache data to the redux store
The problem actually occurs when we execute recordOptimisticTransaction.
During an optimistic transaction on the offline-cache, the data object
that backs up the cache is swapped with a recordingCache. This object is
responsible for recording the data diff for the optimistic transaction.
So the optimistic results are written to the recordingCache which keeps
a data and recordedData set (the calculated diff). Currently after the
write, the offline-cache then extracts all the data (from the data set)
and sends it to the redux store. Our subscription is triggered and we
call replace with the new data which calls restore on the
recordingCache. So this is where the actual “corruption” happens.
recordedData is returned as the patch, the transaction ends and the
recordingCache is replaced by the original ObjectCache object. When our
mutation result is returned from the server and we call extract to save
it to the redux store we get the data + the patch.