Inconsistent store (#43)
* Preserve the observer associated with an existing topic

* Update CHANGELOG.md

* Store mutation update function in AASContext and restore when enqueing
for effect

When calling `enqueueMutation` in `offline-link.js`, the operation
object created by `QueryManager` does not contain the update function.
See `buildOperationForLink`. To preserve the update function, we save it
in `AASContext` which is included (in `context`) when building the
operation

* Stop restoring apollo cache after the appsync store has been rehydrated

Constantly restoring apollo cache from store after writing causes issues
with optimistic responses. The patches are not pruned and ultimately the
appsync store does not end up with the correct data and the apollo cache
is 'corrupted'. Easily reproducible by creating a mutation with a
pessimistic boolean flag on optimistic response, no update function (as
to not update an existing query) and no running subscription.

* Stop persisting recordingCache data to the redux store

The problem actually occurs when we execute recordOptimisticTransaction.
During an optimistic transaction on the offline-cache, the data object
that backs up the cache is swapped with a recordingCache. This object is
responsible for recording the data diff for the optimistic transaction.
So the optimistic results are written to the recordingCache which keeps
a data and recordedData set (the calculated diff). Currently after the
write, the offline-cache then extracts all the data (from the data set)
and sends it to the redux store. Our subscription is triggered and we
call replace with the new data which calls restore on the
recordingCache. So this is where the actual “corruption” happens.
recordedData is returned as the patch, the transaction ends and the
recordingCache is replaced by the original ObjectCache object. When our
mutation result is returned from the server and we call extract to save
it to the redux store we get the data + the patch.

* Update README.md

* Update offline-cache.js

* Update CHANGELOG.md


Fixes #56 
Fixes #59