Delete observer associated with a topic when cleanup function is called
Multiple topics can be associated with a client. This happens when
multiple subscriptions are live on a single API. In that case,
clientTopics holds client => [topics]. When the cleanup function is
called, we should look for a topic within an existing list of topics
associated with the client to identify it. Doing this allows us to find
the correct <topic, client> pair which allows us to unsubsribe the
client from the topic and remove the observer tied to the topic. Leaving
a "completed" observer in `topicObserver` will lead to it being reused
if we cancel a subscription and then switch back to it. However since
the observer has been "completed", its `next` function would never be
called.

In `unsubscribeFromTopic` function, we should not delete a `client` key if the
list of subscribed topics is empty: the client has not been disconnected
yet. The client will be disconnected when `disconnectAll` is called the
next time `request` is invoked (next subscription setup).